Create another component for options variables
