+++ src/components/MinimizeView/MinimizeView.tsx
@@ -1,9 +1,9 @@
-import React, { useCallback, useMemo } from 'react';
-import { EventBus, SelectableValue } from '@grafana/data';
-import { Alert, InlineField, Select, useTheme2 } from '@grafana/ui';
+import React from 'react';
+import { EventBus } from '@grafana/data';
+import { Alert, useTheme2 } from '@grafana/ui';
 import { MinimizeViewOptions } from '../../types';
 import { useRuntimeVariables } from '../../hooks';
-import { selectVariableValues } from '../../utils';
+import { OptionsVariable } from '../OptionsVariable';
 import { Styles } from './styles';
 
 /**
@@ -37,65 +37,6 @@ export const MinimizeView: React.FC<Props> = ({ options: { variable: variableNam
   const { variable } = useRuntimeVariables(eventBus, variableName || '');
 
   /**
-   * Current values
-   */
-  const values = useMemo(() => {
-    if (!variable?.options) {
-      return [];
-    }
-
-    return variable.options.filter((option) => option.selected).map((option) => option.value);
-  }, [variable]);
-
-  /**
-   * On Change
-   */
-  const onChange = useCallback(
-    (value: Array<SelectableValue<string>> | SelectableValue<string>) => {
-      const options = Array.isArray(value) ? value : [value];
-      const updatedValues = options
-        .filter((option) => option.value !== undefined)
-        .map(({ value }) => value) as string[];
-
-      /**
-       * Deselect values
-       */
-      if (values.length > updatedValues.length) {
-        /**
-         * Select all
-         */
-        if (updatedValues.length === 0 && variable?.multi && variable.includeAll) {
-          selectVariableValues(['all'], variable);
-          return;
-        }
-
-        const removedValues = values.filter((value) => !updatedValues.includes(value));
-        selectVariableValues(removedValues, variable);
-        return;
-      }
-
-      /**
-       * Select Values
-       */
-      selectVariableValues(updatedValues, variable);
-    },
-    [values, variable]
-  );
-
-  /**
-   * Options
-   */
-  const options = useMemo(() => {
-    if (!variable) {
-      return [];
-    }
-    return variable.options.map((option) => ({
-      label: option.text,
-      value: option.value,
-    }));
-  }, [variable]);
-
-  /**
    * No variable selected
    */
   if (!variable) {
@@ -108,9 +49,7 @@ export const MinimizeView: React.FC<Props> = ({ options: { variable: variableNam
 
   return (
     <div className={styles.root}>
-      <InlineField label={variableName}>
-        <Select onChange={onChange} options={options} isMulti={variable?.multi} value={values} />
-      </InlineField>
+      {(variable.type === 'query' || variable.type === 'custom') && <OptionsVariable variable={variable} />}
     </div>
   );
 };
