+++ src/components/MinimizeView/styles.ts
@@ -8,6 +8,7 @@ export const Styles = (theme: GrafanaTheme2) => {
   return {
     root: css`
       padding: ${theme.spacing(1)};
+      overflow: auto;
     `,
   };
 };
