+++ src/components/OptionsVariable/index.ts
@@ -1,0 +1,1 @@
+export * from './OptionsVariable';
